incorporated changes for auth and board features
