--- src/model/state/jobs/selector.js
@@ -1,3 +1,0 @@
-export const selectJob = state => state.job;
-
-export const selectJobList = state => state.jobList;
