+++ src/model/state/selectors.js
@@ -1,2 +1,6 @@
 
 export const selectHistory = state => state.history;
+
+export const selectJob = state => state.job;
+
+export const selectJobList = state => state.jobList;
