+++ src/view/dashboard/containers/Header.jsx
@@ -8,13 +8,13 @@ const Header = ( ) => {
     const token = window.localStorage.getItem('kondotoken');
     const payload = jwt.decode(token);
     const location = useLocation();
-    const icon = require('../_images/logo.png');
+    const icon = require('../../../_images/logo.png');
 
     console.log(payload)
 
     return (
-        <div className='header'>
-            <div className='logo'>
+        <div className='header' style={{display: 'flex'}}>
+            <div className='logo' style={{display: 'flex'}}>
             <img src={icon} alt='KondoBoard Logo'/>
             <p>KondoBoard</p>
             </div>
