+++ src/view/dashboard/containers/NotFound.jsx
@@ -1,7 +1,7 @@
 import React, {useState} from 'react';
 import { Redirect } from 'react-router-dom';
 import { useSelector } from 'react-redux';
-import { selectHistory } from '../model/state/selectors'; 
+import { selectHistory } from '../../../model/state/selectors.js'; 
 
 const NotFound = () => {
