+++ src/view/dashboard/jobsearchcomponents/JobList.jsx
@@ -2,7 +2,7 @@ import React, {useEffect} from 'react';
 import {useSelector} from 'react-redux';
 
 import JobController from '../../../controllers/JobController.js';
-import {selectJobList} from '../../../model/state/jobs/selector.js';
+import {selectJobList} from '../../../model/state/selectors.js';
 
 import Job from './Job.jsx';
