--- src/view/dashboard/list/ListView.jsx
@@ -1,14 +1,0 @@
-// import React from 'react';
-
-// import Header from '../../headers/Header';
-
-// const ListView = () => {
-//     return(
-//         <div>
-//             <Header />
-//             <h1>List</h1>
-//         </div>
-//     )
-// }
-
-// export default ListView;
